Fix s:WrapSelection for trailing multibyte characters
If a selection ends in a multibyte character and the user calls
s:WrapSelection (via PareditWrap for instance), only the first byte of
the last character is wrapped.

e.g. (^ is the insertion caret, , is g:paredit_leader)

    my-λ^

User presses ,w

    (my-<ce>)<bb>

Where <xx> are the literal hex values for the character λ in unicode.

We can hack around this by using matchstr() to return the last _char_ in
the selection, then use strlen() to determine the new byte offset.

There are likely other multibyte errors such as this one, but this is
the only one I have noticed.